# Optimise le temps de chargement de la liste des élèves