--- a/tests/test_eleve_form.rb
+++ b/tests/test_eleve_form.rb
@@ -677,6 +677,16 @@ class EleveFormTest < Test::Unit::TestCase
   #   assert doc.css("select[name='classes'] option").collect(&:text).include? "4EME ULIS"
   # end
 
+  def test_liste_des_eleves
+    eleve = Eleve.find_by(identifiant: 2)
+
+    post '/agent', identifiant: 'pierre', mot_de_passe: 'demaulmont'
+    get '/agent/liste_des_eleves'
+    doc = Nokogiri::HTML(last_response.body)
+    assert_equal "Edith", doc.css("##{eleve.dossier_eleve.id} td:nth-child(2)").text.strip
+    assert_equal "Piaf", doc.css("##{eleve.dossier_eleve.id} td:nth-child(3)").text.strip
+  end
+
   def test_affiche_changement_adresse_liste_eleves
     # Si on a un changement d'adresse
     eleve = Eleve.find_by(identifiant: 2)