Dans l'environnement de développement, on passe de 3.9s à 2.3s pour le chargement de la liste des élèves; le temps effectif de rendu de la page est de 150ms, le rendu d'une page vide prend 2.25s. Le résultat réel est à vérifier sur staging mais cette PR devrait rendre le chargement de la page presque instantané.

La version optimisée est plus susceptible de régressions, puisqu'on interroge directement des jointures SQL plutôt que des objets du modèle dont les attributs (et leur unicité) sont garantis. Ainsi `etat` de la pièce jointe peut être ambigu avec `etat` du dossier élève. Des tests unitaires complémentaires s'imposent donc sur la liste des élèves, cette PR en ajoute 2.

Un changement de comportement par rapport à la version antérieure: on effectue la jointure seulement sur le responsable légal de priorité 1, donc on n'affichera un picto enveloppe que si ce responsable a déclaré son adresse mail, et non plus l'un des deux.